+++ process-instance-migration/src/test/resources/application.yaml
@@ -1,5 +1,3 @@
-server.port: 18080
-
 camunda:
   bpm:
     auto-deployment-enabled: false
@@ -12,7 +10,3 @@ operate:
         username: demo
         password: demo
 
-camunda7:
-  client:
-    base-url: http://localhost:18080/engine-rest
-    check-on-init: false
